Made separated files with functions and class
